Delete Endgame from everything. Put text into GameWords. PUpdated tests, GameTest still not working properly.
