--- lib/endgame.rb
@@ -1,106 +1,0 @@
-
-class Endgame
-
-  def initialize(endgame)
-    @endgame = endgame
-  end
-
-
-  def player_wins
-    puts "*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*"
-    puts "YOU. SUNK. OUR. BATTLESHIPS!"
-    puts "*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*"
-    puts "Congratulation #{@user_name}"
-    puts "You've bested our best commander,"
-    puts "And for that, we're awarding you our"
-    puts "biggest trophy! Are you readyyy?"
-    sleep(8)
-    puts "        ...................................................................................................."
-    puts "    ...................................................................................................."
-    puts "    ....... ..DMDN=$.................................................................. D+N~N.. .  ......"
-    puts "    ...... .??=I7.+$N=. .. ..........................................................,.M$IN=$?M........."
-    puts "    .......$+=~M:$~N+8...............................................................:=N,~IMNNMO ......."
-    puts "    .......+.7NI..Z:MM... ......................~,,,.................................ZMO7..D87O7........"
-    puts "    ..... ..,N,. .~I,+......=IDDONMMOZI:DNMNNDDO$$77I+?IIII$ODNDDNN,IZ$88OZO7..... ..~7I:...8++, ......."
-    puts "    ..... .=Z8I~....D$.....?+IO?=MMM8?:,DMNNDD8Z$$77777?II77$OODNMM,:7ZNNN=ZD88D....87O....,ONO, ......."
-    puts "    ...... .NI+I.....,O,  .I?+Z8MMDO88OO8.......~$7Z77ZOZO...,,...IDD88OZ?I87I+,...=NN.....$:8D........."
-    puts "    ...... .:N,=~....I,D...+??Z8MMMM8MMDD.......NMMMMMMMMMM7......DMMMMMII77$?+?...Z8$.....=:$$........."
-    puts "    ...... ..~,O:.. ..Z=NO$~?+$O8888D888O....,+ZNMMMMMMMMMMND88?.,?$ZOOOZII+OZ+?IO+??. ...N.:Z.. ......."
-    puts "    ..........$:,O..  . .  =++7$ZZZZND7?~,,.,,,,:~~=======~~:,,,,:~+?I7$+I:8O7++.=..... .O~,I. ........."
-    puts "    .... .... .~+OI.. ... ..+~?I777I??+==~:::::~~=$=I?+I+==~~::::~=++??NIIIZI++. .......,=.+............"
-    puts "    .............D+D .......=7???+ZO?+++==~~~~~=~~ZIIN$I???=~~~~==++++??77ZZ+I=.........DN~. ..........."
-    puts "    ..... .. .....$I?.....  .+7$ZZ$?O+++=======~~ZZIZO7O?7?======++++??7II?+++...... ..DD..............."
-    puts "    ...... ........I7+........+?7N=87?+++=====++8D?IO8ZI7$=+=====++++I$$I+++7=.......DN=................"
-    puts "    ......... .. ...II+.......=?D?:~D?I++======+?Z8I7877OI~+====++++?I?I?+=++.......8N+................."
-    puts "    ........ .. ......~I?..  ..=$MNNZ77?++++===8+?$DZZ7?7==:==++++?8ZOI?++?=......:IN... ... ..........."
-    puts "    ....... ...........++D..... ,?+$$DZ?I???+===.:??7Z???.~=++++?IN77I?++?=.....,OZ~...................."
-    puts "    ...... .. ...........~,=... ..7++?II7?NI?+++====$87~=++++++?INNN?+++I. ....I,:......................"
-    puts "    ......................==II...$$Z++??IONZ7??++++++++++++++???III?+++7$8...O?=7......................."
-    puts "    ....... ................,Z+::ZI..?=+?II77I??IM8$7II7Z??????????++Z..D7=7=+=... ....................."
-    puts "    ...........................=+?,....:7+?IIIIIIZ8+7Z?8OZI???????I?.....I=~. .... ....................."
-    puts "    ......... ....................II,.....=++?I??II7+OZ7II????++=......O=:............. ................"
-    puts "    ................... ............7+ ...... ,+O8ZI$7=8I$+=:. ......7=. ..............................."
-    puts "    ...............................I=$..........,78$O78OI............Z8I ..............................."
-    puts "    .................................=?= ......~I$D+OIOO7I ....... ,+I..... ............... ............"
-    puts "    ... ....................... ........... ..:ODDZ7NN+DNI+................ ............................"
-    puts "    ..........................................+7,8 ~MN..N$+............................................."
-    puts "    ..........................................?ODZ?7ZO$ZOO7= ..........................................."
-    puts "    ...........................................?7?7II?$???+. ...... ... ................................"
-    puts "    ............................................=I?O??+?~=........................... .................."
-    puts "    .............................................+~====7+..............................................."
-    puts "    .............................................++===?=........... ...................................."
-    puts "    .............................................+~~~~~I.... ............ .............................."
-    puts "    .............................................I7+==??................................................"
-    puts "    .............................................Z~.NI,$........... .......................... ........."
-    puts "    ............................................=~=+NN$,+..... .. ... . ...... ............... ...... .."
-    puts "    ............................................I~DODD=?I~... ....... ....... ....  ....  .. ..........."
-    puts "    ...........................................IZ.N..,.~OI=............................................."
-    puts "    .........................................+?ZNOZ??I?I?D+=........... ................................"
-    puts "    .........................................=Z$8I=$$$II8Z$= ..........................................."
-    puts "    .........................................+NNN..DMM..NMD=....... .. ... .. ........ .... ............"
-    puts "    ........................................=?ZDN..DMN..8O7+=. ........................................."
-    puts "    .......................................+IIO8NN......DD$O?= ........................................."
-    puts "    .......................................?8$INDNNMMMNNDD$78+~...... .................................."
-    puts "    ......................................=I8NZD. .NMM...ON+8?=........................................."
-    puts "    ......................................+I$OM$ .$NMMD..7MO?++. ......................................."
-    puts "    .....................................=?NZ?ZN8OODNDZ~DDN8$8+=........................................"
-    puts "    .....................................+7777N...NNMMD...NN7ZI=. ......................................"
-    puts "    ...................................I++?O$DI?..8MMMN...=IND+==......................................."
-    puts "    ...................................I$$?~$Z8...OMMMN...??+,O+Z..... .. ...  . . ..... .....  ........"
-    puts "    ...................................DDDDDDND888DNDDD888D88OOO8$......................................"
-    puts "    ...................................NMMMNNNNNNNNMNMNNN8MNNMMMND ....................................."
-    puts "    ...................................NMN7==================+ONM8 ....................................."
-    puts "    ...................................NMD====?+?==+?=?~?=?====MND ....................................."
-    puts "    ...................................NMD~~==~~==~============MNO ....................................."
-    puts "    ...................................NMO~~~~~=~~=??=~====~=~=DN8......................................"
-    puts "    ...................................MM$~~~~~I=~~~~+~==+~~~==NDO ... ................................."
-    puts "    ...................................MM8=~~~~~~====~~=~=~~~~~NNO ........................... ........."
-    puts "    ...................................MMMM~::::::::::::~:::::8MNO ....................................."
-    puts "    ...................................NNMMMMMMMMMMMMMMMMMMMMNMMNZ ....................................."
-    puts "    ..................................... ...  .. ........................... ............. ............"
-    puts "    ...................................................................................................."
-    puts "    ................................................ .. ..... ................. ........... ............"
-    puts "    .... ...................................................... ..................... .................."
-    sleep(10)
-    puts "COME BACK!  ZOOM BACK IN!"
-    sleep(5)
-    puts "We think you're great, and would"
-    puts "love to play again. Would you?"
-  end
-
-  def player_loses
-    puts "*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*"
-    puts "AWwww dang it. You lost :("
-    puts "*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*"
-    puts "Well I have to say you actually did get"
-    puts "pretty close. If George hadn't lucked"
-    puts "out on that last one, you would have"
-    puts "had me for sure. What do you say?"
-    puts "One more game?"
-    sleep(5)
-  end
-end
-
-# the_end = Endgame.new("go")
-# the_end.player_wins
-#
