+++ runner.rb
@@ -1,9 +1,10 @@
-require './lib/turn'
 require './lib/ship'
 require './lib/cell'
 require './lib/board'
 require './lib/tutorial'
-require './lib/endgame'
+require './lib/game_words'
+require './lib/turn'
+require './lib/game'
 
 game = Game.new
 game.start
